requirements.txt
Flask>=2.0
python-telegram-bot>=13.0
requests>=2.0
flask-cors>=3.0
pytest>=7.0